operations_django/requirements.txt
django
djangorestframework
uwsgi
psycopg
